==> Rmd/01-gather-data.Rmd <==
---
title: "Gathering Data"
author: "Eric C. Anderson"
date: "November 18, 2014"
output: html_document
---

```{r, echo=FALSE, include=FALSE}
library(knitr)
opts_knit$set(root.dir = '..')
```

```{r, echo=FALSE, include=FALSE}
# load libraries, etc
library(ggplot2)
library(lubridate)
library(stringr)
library(dplyr)
library(reshape2)
```
Here I document how I went about gathering data from RMIS.  Mostly it is just
a record of the Queries used and where I put the resulting data files.


## Releases

I wanted to get all releases, so I followed this link http://www.rmis.org/cgi-bin/queryfrm.mpl?Table=all_releases&Version=4.1
and then choose all possible years and asked for it as a csv file.

```
The query criteria used to produce this report was:

WHERE brood_year IN (1950,1951,1952,1953,1954,1955,1956,1957,1958,1959,1960,1961,1962,1963,1964,1965,1966,1967,1968,1969,1970,1971,1972,1973,1974,1975,1976,1977,1978,1979,1980,1981,1982,1983,1984,1985,1986,1987,1988,1989,1990,1991,1992,1993,1994,1995,1996,1997,1998,1999,2000,2001,2002,2003,2004,2005,2006,2007,2008,2009,2010,2011,2012,2013,2014)
```

Once I downloaded it and unzipped it I did this:
```{r, eval=FALSE}
x <- read.csv("~/Downloads/CSV5964.TXT", stringsAsFactors = FALSE, na.strings = "")
x <- read.csv(unz("~/Downloads/CSV5964.zip", filename = "CSV5964.TXT"), stringsAsFactors = FALSE, na.strings = "")
x <- tbl_df(x)
saveRDS(x, compress = "xz", file = "data/releases.rds") 
```



## Recovery Data

After looking over the release data I decided that it would be best to focus on recent
years for our comparison.  In particular, recoveries made grom 2002 to 2012 seemed appropriate.
So, I tried that first for Chinook.  That would have returned 1.4 million rows which is more
tham RMIS wants to hand out.  It would be preferable if RMIS didn't throttle that.
So, I tried breaking it into smaller queries, but it is 
necessary to break it down year by year.  Holy Bagumba!  So that is what I did.  In fact, for 2012
I had to break it down even further by "Sample Type" and get it in two installments.

Fortunately, the 12 emails resulting from those queries got put in a single thread in gmail and I was 
able to copy them all into a text file which I have saved in 
`data_gathering/rmpc_recovery_query_emails.txt`.  We can use R's text processing and system call 
abilities, and its ability to download files in order to stitch them into a single csv that
we can then read in.  In fact, let's use the `stringr` package.
```{r, eval=FALSE}
email <- readLines("data_gathering/rmpc_recovery_query_emails.txt")

# get the URLs to download the files
ziplinks <- email[which(str_detect(email, "PKZIP")) + 1] %>% str_trim()

# cycle over those, download, unzip, and read into data frames.
# download and unzip to a temporary directory so they don't bomb our repo
td <- tempdir()
dfs <- lapply(ziplinks, function(x) {
  fn <- str_replace(x, "http:.*/", "") 
  fnp <- file.path(td, fn)  # path to download to
  download.file(x, fnp)
  unzip(fnp, exdir = td)
  read.csv(str_replace(fnp, "zip$", "TXT"), stringsAsFactors = FALSE, na.strings = "")
  })

recoveries <- tbl_df(do.call(rbind, dfs))
```

Having gotten all that, let us save it compress it.
```{r, eval=FALSE}
saveRDS(recoveries, compress = "xz", file = "data/chinook_recoveries.rds")
```
It gets whittled down to only about 6.6 MB.  Cool.

Here are the queries for all the data that I got:
```{r}
email <- readLines("data_gathering/rmpc_recovery_query_emails.txt")
cat(email[str_detect(email, "^The query criteria|^AND|^WHERE")], sep = "\n")
```


## Catch Sample data
The query was:
```
WHERE catch_year IN (2002,2003,2004,2005,2006,2007,2008,2009,2010,2011,2012)
```

And I processed the results like this:
```{r, eval=FALSE}
cs <- read.csv("~/Downloads/CSV10233.TXT", stringsAsFactors = FALSE, na.strings = "")
catch_sample <- tbl_df(cs)
saveRDS(catch_sample, compress = "xz", file = "data/catch_sample.rds")
```


## Location data

I figured it would be feasible to just download the entire location data base.  The query that seemed
to make that possible produced this in the email from RMIS:
```
WHERE state IN ('AK','BC','CA','CO','FO','HS','ID','MI','MN','MT','ND','NE','OR','WA','WI','WY')
```
That's fine.  I did that query on November 4, 2014.  It is a csv file and I have placed it into:
`./data/locations.csv`
