==> Rmd/02-01-explore-release-data.Rmd <==
---
title: "Exploring Release Data"
author: "Eric C. Anderson"
date: "November 18, 2014"
output:
  html_document:
    fig_width: 10
    fig_height: 8
    toc: yes
---

```{r, echo=FALSE, include=FALSE}
library(knitr)
opts_knit$set(root.dir = '..')
```

```{r, echo=FALSE, include=FALSE}
# load libraries, etc
library(ggplot2)
library(lubridate)
library(dplyr)
library(reshape2)
source("R/rmis_cleaning.R")
```


## Initial Maneuvers

Here we look at some properties to make sure that we got everything we wanted:
```{r}
# load it in
releases <- readRDS("data/releases.rds")
dim(releases)
releases$species <- int_to_factor(releases$species, species_levels())
```

### Sizes of tag-associated release groups

Let's quickly compile some information on how many fish likely
produced each of these release groups.
```{r}
relcnts <-releases %>% filter(record_code == "T", brood_year > 2000) %>% select(reporting_agency, ends_with("count"), brood_year, tag_code_or_release_id)

# change NAs in the counts to 0
relcnts[, 2:5][is.na(relcnts[,2:5])] <- 0

femcnts <- relcnts %>% 
  mutate(tots = cwt_1st_mark_count + cwt_2nd_mark_count + non_cwt_1st_mark_count + non_cwt_2nd_mark_count) %>%
  select(tag_code_or_release_id, reporting_agency, brood_year, tots) %>%
  mutate(tot_fem = tots / 4000) %>%
  mutate(fem_cats = cut(tot_fem, breaks = c(0, 10, 25, 50, 100, 250, 500, 1000, 10^4)))

# how many total fish released from each "number of females category"
femcnts %>%
  group_by(fem_cats) %>%
  summarise(tot_rel = sum(tots)) %>%
  mutate(fract = tot_rel / sum(tot_rel))
  
```


### Numbers of tagged vs untagged fish etc.
We will start off with some simple investigations of the number of fish 
that are tagged and not tagged.  These numbers are stored in four columns
named `r paste(names(releases)[grepl("_count$", names(releases))], collapse=", ")`.

So, we can sum those up and plot them by year
```{r}
# count em and reshape
count_cols <- names(releases)[grepl("_count$", names(releases))]
counts <- releases %>%
  group_by(brood_year, species, record_code) %>%
  summarise_each_(funs(sum(., na.rm = TRUE)), count_cols) %>%
  melt(id.vars = c("species", "brood_year", "record_code"))

# plot em
ggplot(counts, aes(x = brood_year, y = value, color =  variable, linetype = record_code)) + 
  geom_line() + 
  facet_wrap(~ species, ncol = 3)
```

That is very interesting. It shows that there are a lot of untagged pink and chums
released.  It also shows that the tagging fraction of chinook appears to be
the highest amongst all the species.

It is important to point out that the solid lines are the //non-associated
releases//---i.e. releases of hatchery fish that do not include any tagged
fish.  And the dashed lines are the releases that include fish with tags.  

Let us now look closely at Chinook and Coho:
```{r}
counts %>%
  filter(species %in% c("Chinook", "Coho")) %>%
  ggplot(aes(x = brood_year, y = value, color =  variable, linetype = record_code, shape = record_code)) + 
  geom_point() + geom_line() +
  facet_wrap(~ species, ncol = 2, scales = "free")

```

Again, this is very interesting.  One of the things that I notice is that
there are a lot of non-cwt-2nd-mark-count fish in the late '90s.  I wonder
if the non-cwt-first-mark fish in those cases are totally unmarked, and the
second mark in that case is the ad-clip.

### What years shall we focus on?
In our coastwide simulation, I think it would be good to compare how PBT would have fared versus
CWTs for a reasonably long period of time, and obviously it will be worthwhile to have it be a 
fairly recent block of time.  It is pretty clear that the release data after 2012 is not really there,
which makes me wonder if the recovery data might be incomplete too.  So, I propose this:

* __Focus on recoveries from 2002 to 2012__
* __Hence, we will want to deal with releases from 1993 to 2010__

So, let's see what those look like:
```{r}
counts %>%
  filter(species %in% c("Chinook", "Coho"), brood_year >= 1993 & brood_year <= 2010 ) %>%
  ggplot(aes(x = brood_year, y = value, color =  variable, linetype = record_code, shape = record_code)) + 
  geom_point() + geom_line() +
  facet_wrap(~ species, ncol = 2, scales = "free")
```

That looks pretty darn sensible to me.  So, moving forward we will want to filter on these.

### What sorts of marks are being used?

The PSC has an elaborate coding system for combinations of different marks.  Unfortunately it seems that
they only have it recorded in a PDF file.  It seems that they should have it in a flat text file
so that one could do something with it easily.  I have tried my best to faithfull translate the marks
into a factor.  Let's do that here:
```{r}
markls <- releases %>% 
  mutate_each_(funs = funs(mark_to_string), names(releases)[grepl("_mark$", names(releases))])
```

Great, now I want to look at upon how many fish each of these is used for the four catetegories of marks (cwt, non) x (1st, 2nd).  We may as well
limit ourselves to the species and years we are interested in.  In fact, let's just do this for Chinook right now.

```{r}
chinook18 <- markls %>% 
  filter(species %in% c("Chinook"), brood_year >= 1993 & brood_year <= 2010 )
```

```{r}
marksums <- chinook18 %>% 
  group_by(record_code, cwt_1st_mark,cwt_2nd_mark, non_cwt_1st_mark, non_cwt_2nd_mark) %>% 
  summarise_each(funs(sum(., na.rm = TRUE)), cwt_1st_mark_count, cwt_2nd_mark_count, non_cwt_1st_mark_count, non_cwt_2nd_mark_count) %>%
  ungroup()

totals <- marksums %>% select(ends_with("_count")) %>% rowSums( na.rm = TRUE)

# tack on totals, percentages, and then sort
marksums <- cbind(marksums, total = totals) %>% 
  arrange(desc(total)) %>%
  mutate(percentage = 100 * total / sum(total), cumul_perc = cumsum(percentage))

```

That is awfully hard to look at in R.  So, it will get saved as a csv file:
```{r}
write.csv(marksums, file = "marktypes.csv")
```

### Super rough tagging costs calculations

Looking at the table that got output above makes me start wondering how much it would
cost to tag every ad-clipped fish (from associated release groups, or not) using PBT,
and it would be good to compare that to the CWT cost.

We will want to account for DIT tagging pecularities, too.  

So, first off, let's get the number of CWT tagged fish for every release groups, regardless of the kind
of mark that they got, and we can multiply that by 13 cents for the price of every wire tag.
```{r}
cwts <- chinook18 %>%
  group_by(brood_year) %>%
  summarise_each(funs(sum(., na.rm = TRUE)), cwt_1st_mark_count, cwt_2nd_mark_count) %>%
  mutate(num_cwt = cwt_1st_mark_count + cwt_2nd_mark_count) %>% 
  mutate(cwt_dollars = num_cwt * 0.13) %>% 
  select(brood_year, num_cwt, cwt_dollars)
```

OK, so, by 2010, we are in the 7.3 million dollar range just for the price of the coded wire.  

Next, we are going to calculate the number of fish that have ad-clips but do not have coded wire
tags.  We can do this both for unassociated releases and associated releases.  Our strategy is to
count the number of non-cwt fish with an adclip of any sort (wether it is a first or second mark)
and then sum those and break them down by year.
```{r}
# first get number of adclipped fish with no CWTs
non_cwts_adclip_vec <- chinook18 %>%
  mutate(non_cwt_ad1 = grepl("^Adclip", non_cwt_1st_mark) * non_cwt_1st_mark_count,
         non_cwt_ad2 = grepl("^Adclip", non_cwt_2nd_mark) * non_cwt_2nd_mark_count) %>%
  select(non_cwt_ad1, non_cwt_ad2) %>% rowSums(na.rm = TRUE)

non_cwts <- tbl_df(cbind(chinook18, non_cwts_adclip_vec = non_cwts_adclip_vec)) %>% 
  group_by(record_code, brood_year) %>%
  summarize(non_cwt_adclip = sum(non_cwts_adclip_vec)) %>%
  dcast(brood_year ~ ...)
non_cwts
```

OK, N is the column of non-associated releases and T is column of tagged releases.  Let's compute other costs on that.  We'll assume
a female chinook produces 4K smolts and that the sex ratio in the hathery is even so that you get 2K smolts for every
spawner genotyped.  And, let's do $20 per genotype.
```{r}
inner_join(cwts, non_cwts) %>%
  mutate(pbt_dollars_cwt_num = num_cwt * 20 / 2000,
         pbt_dollars_cwt_N_T = (num_cwt + N + T) * 20 / 2000)
```


### Can we tidy this up?

So, having the mark type and then the numbers in those awful columns like `non_cwt_1st_mark_count`
is kinda lame.  It feels like I'm fighting it the whole way.  I am going to make it tidy.

Making it tidy takes a bit of wrangling.  We basically want to think of something that
is designed such that if tag_code `12AA5` had 5321 fish that had a cwt and were marked
with an adclip (as part of the two possible differently-marked cwt groups) and 30000 fish
that had no cwt and were marked with adclip too, would show up like this:
```{r, eval=FALSE}
tag_code      with_cwt      mark_series   mark_type   count
12AA5              yes                1      adclip    5321
12AA5               no                1      adclip   30000
```

I don't know how to do that with reshape2, but I can write some lapply foo and then throw
an inner join into it that will work out:
```{r}
tmp <- chinook18 %>%
  select(tag_code_or_release_id,cwt_1st_mark:non_cwt_2nd_mark_count)

# now a function that grabs stuff out as need be and makes a small data frame of it:
grab_it <- function(x, C = "cwt", S = "1st") {
  with_cwt_v <- unname(c(cwt = "yes", non_cwt = "no")[C])
  mark_series_v <- as.numeric(substr(S, 1, 1))

  Mark_col <- paste(C, S, "mark", sep = "_")
  Count_col <- paste(C, S, "mark_count", sep = "_")
  
  tt <- data.frame( tag_code_or_release_id = x[["tag_code_or_release_id"]], 
              with_cwt = with_cwt_v,
              mark_series = mark_series_v,
              mark_type = x[[Mark_col]], 
              count = x[[Count_col]],
              stringsAsFactors =  FALSE)
  tt
}

# then lapply and rbind and drop rows with NAs
boing <- c(lapply(c("cwt", "non_cwt"), function(z) grab_it(tmp, z, "1st")),
           lapply(c("cwt", "non_cwt"), function(z) grab_it(tmp, z, "2nd")))
finally <- na.omit(do.call(rbind, boing))

# and now we can inner_join that 
tidy18 <- tbl_df(inner_join(finally, select(chinook18, -(cwt_1st_mark:non_cwt_2nd_mark_count))))
```

Now that we have gone through that crap there, it should be delightfully easy to get the 
numbers that we want.  For example, the total number of fish with cwts by year:
```{r}
tidy18 %>% 
  group_by(with_cwt, brood_year) %>%
  summarize(sum(count, na.rm = TRUE))
```

And, we can recreate the plot of chinook and coho CWT and non-CWT numbers like this:
```{r}
tidy18 %>% 
  group_by(brood_year, record_code, with_cwt, mark_series) %>%
  summarize(number_of_fish = sum(count, na.rm = TRUE)) %>%
  ggplot(aes(x = brood_year,
                   y = number_of_fish,
                   colour = with_cwt,
                   linetype = record_code,
                   shape = factor(mark_series))) +
  geom_line() +
  geom_point()
```

OK, so that is nice to have things tidied up.  Now, let's get the total number of ad-clipped fish
that are either carrying CWTs or not:
```{r}
tidy18 %>%
  filter(grepl("^Adclip", mark_type) ) %>%
  group_by(brood_year, with_cwt, record_code) %>% 
  summarize(num_fish = sum(count, na.rm = TRUE))
```
